+++ lib/src/interpreter.dart
@@ -10,7 +10,6 @@ import 'class.dart';
 import 'function.dart';
 import 'lexer.dart';
 import 'parser.dart';
-import 'resolver.dart';
 import 'lexicon.dart';
 
 typedef ReadFileMethod = dynamic Function(String filepath);
@@ -19,7 +18,6 @@ String syncReadFileMethod(String filapath) => File(filapath).readAsStringSync();
 
 /// 负责对语句列表进行最终解释执行
 class Interpreter implements ExprVisitor, StmtVisitor {
-  final String workingDirectory;
   final bool debugMode;
   final ReadFileMethod readFileMethod;
 
@@ -39,11 +37,12 @@ class Interpreter implements ExprVisitor, StmtVisitor {
   HT_Namespace curContext;
   String _curFileName;
   String get curFileName => _curFileName;
+  String _curDirectory;
+  String get curDirectory => _curDirectory;
 
   dynamic _curStmtValue;
 
   Interpreter({
-    this.workingDirectory = 'script/',
     this.debugMode = false,
     this.readFileMethod = defaultReadFileMethod,
   }) {
@@ -55,15 +54,17 @@ class Interpreter implements ExprVisitor, StmtVisitor {
   dynamic eval(
     String content, {
     String fileName,
+    String libName = HT_Lexicon.globals,
     HT_Namespace context,
     ParseStyle style = ParseStyle.library,
     String invokeFunc,
-    List<dynamic> args,
+    Map<String, dynamic> args,
   }) {
+    _curFileName = fileName;
+    _curFileName ??= '__anonymousScript' + (Lexer.fileIndex++).toString();
+
     curContext = context ?? _globals;
-    final tokens = Lexer().lex(content);
-    final statements = Parser(this).parse(tokens, fileName: fileName, style: style);
-    Resolver(this).resolve(statements, fileName: fileName);
+    final statements = Lexer(this, content, fileName: _curFileName).lex().parse(style: style).resolve(libName: libName);
     for (final stmt in statements) {
       _curStmtValue = evaluateStmt(stmt);
     }
@@ -79,13 +80,16 @@ class Interpreter implements ExprVisitor, StmtVisitor {
   /// 解析文件
   Future<dynamic> evalf(
     String filepath, {
+    String directory,
     String libName,
     ParseStyle style = ParseStyle.library,
     String invokeFunc,
-    List<dynamic> args,
+    Map<String, dynamic> args,
   }) async {
     final savedFileName = _curFileName;
+    final savedFileDirectory = _curDirectory;
     _curFileName = filepath;
+    _curDirectory = directory ?? File(_curFileName).parent.path;
     dynamic result;
     if (!_evaledFiles.contains(curFileName)) {
       if (debugMode) print('hetu: Loading $filepath...');
@@ -93,27 +97,32 @@ class Interpreter implements ExprVisitor, StmtVisitor {
 
       HT_Namespace library_namespace;
       if ((libName != null) && (libName != HT_Lexicon.globals)) {
-        _globals.define(libName, this, declType: HT_Type.NAMESPACE);
-        library_namespace = HT_Namespace(name: libName, closure: library_namespace);
+        library_namespace = HT_Namespace(name: libName, closure: _globals);
+        _globals.define(libName, this, declType: HT_Type.NAMESPACE, value: library_namespace);
       }
 
       var content = await readFileMethod(_curFileName);
-      result = eval(content.toString(),
+
+      result = eval(content,
           fileName: curFileName, context: library_namespace, style: style, invokeFunc: invokeFunc, args: args);
     }
     _curFileName = savedFileName;
+    _curDirectory = savedFileDirectory;
     return result;
   }
 
   Future<dynamic> evalfSync(
     String filepath, {
+    String directory,
     String libName,
     ParseStyle style = ParseStyle.library,
     String invokeFunc,
-    List<dynamic> args,
+    Map<String, dynamic> args,
   }) async {
     final savedFileName = _curFileName;
+    final savedFileDirectory = _curDirectory;
     _curFileName = filepath;
+    _curDirectory = directory ?? File(_curFileName).path;
     dynamic result;
     if (!_evaledFiles.contains(curFileName)) {
       if (debugMode) print('hetu: Loading $filepath...');
@@ -126,10 +135,11 @@ class Interpreter implements ExprVisitor, StmtVisitor {
       }
 
       var content = syncReadFileMethod(_curFileName);
-      result = eval(content.toString(),
+      result = eval(content,
           fileName: curFileName, context: library_namespace, style: style, invokeFunc: invokeFunc, args: args);
     }
     _curFileName = savedFileName;
+    _curDirectory = savedFileDirectory;
     return result;
   }
 
@@ -216,13 +226,13 @@ class Interpreter implements ExprVisitor, StmtVisitor {
     return _globals.fetch(key, null, null, this, from: _globals.fullName);
   }
 
-  dynamic invoke(String name, {String classname, List<dynamic> args}) {
+  dynamic invoke(String name, {String classname, Map<String, dynamic> args}) {
     HT_Error.clear();
     try {
       if (classname == null) {
         var func = _globals.fetch(name, null, null, this, recursive: false);
         if (func is HT_Function) {
-          return func.call(this, null, null, args ?? []);
+          return func.call(this, null, null, namedArgs: args);
         } else {
           throw HTErr_Undefined(name, null, null, curFileName);
         }
@@ -232,7 +242,7 @@ class Interpreter implements ExprVisitor, StmtVisitor {
           // 只能调用公共函数
           var func = klass.fetch(name, null, null, this, recursive: false);
           if (func is HT_Function) {
-            return func.call(this, null, null, args ?? []);
+            return func.call(this, null, null, namedArgs: args);
           } else {
             throw HTErr_Callable(name, null, null, curFileName);
           }
@@ -293,6 +303,11 @@ class Interpreter implements ExprVisitor, StmtVisitor {
       map[key] = value;
     }
     return map;
+  }
+
+  @override
+  dynamic visitLiteralFunctionExpr(LiteralFunctionExpr expr) {
+    return HT_Function(expr.funcStmt, declContext: curContext);
   }
 
   @override
@@ -423,18 +438,23 @@ class Interpreter implements ExprVisitor, StmtVisitor {
   @override
   dynamic visitCallExpr(CallExpr expr) {
     var callee = evaluateExpr(expr.callee);
-    var args = <dynamic>[];
-    for (final arg in expr.args) {
-      var value = evaluateExpr(arg);
-      args.add(value);
+    var positionedArgs = [];
+    for (var i = 0; i < expr.positionedArgs.length; ++i) {
+      positionedArgs.add(evaluateExpr(expr.positionedArgs[i]));
+    }
+
+    var namedArgs = <String, dynamic>{};
+    for (var name in expr.namedArgs.keys) {
+      namedArgs[name] = evaluateExpr(expr.namedArgs[name]);
     }
 
     if (callee is HT_Function) {
       if (callee.funcStmt.funcType != FuncStmtType.constructor) {
         if (callee.declContext is HT_Instance) {
-          return callee.call(this, expr.line, expr.column, args ?? [], instance: callee.declContext);
+          return callee.call(this, expr.line, expr.column,
+              positionedArgs: positionedArgs, namedArgs: namedArgs, instance: callee.declContext);
         } else {
-          return callee.call(this, expr.line, expr.column, args ?? []);
+          return callee.call(this, expr.line, expr.column, positionedArgs: positionedArgs, namedArgs: namedArgs);
         }
       } else {
         //TODO命名构造函数
@@ -450,7 +470,8 @@ class Interpreter implements ExprVisitor, StmtVisitor {
       //   }
       // }
 
-      return callee.createInstance(this, expr.line, expr.column, curContext, args: args);
+      return callee.createInstance(this, expr.line, expr.column, curContext,
+          positionedArgs: positionedArgs, namedArgs: namedArgs);
     } else {
       throw HTErr_Callable(callee.toString(), expr.callee.line, expr.callee.column, curFileName);
     }
@@ -541,10 +562,7 @@ class Interpreter implements ExprVisitor, StmtVisitor {
   }
 
   @override
-  dynamic visitImportStmt(ImportStmt stmt) {
-    final file_loc = workingDirectory + stmt.location;
-    return evalfSync(file_loc, libName: stmt.nameSpace);
-  }
+  dynamic visitImportStmt(ImportStmt stmt) {}
 
   @override
   dynamic visitVarDeclStmt(VarDeclStmt stmt) {
@@ -668,6 +686,28 @@ class Interpreter implements ExprVisitor, StmtVisitor {
     curContext.define(stmt.name, this,
         declType: HT_Type.CLASS, line: stmt.keyword.line, column: stmt.keyword.column, value: klass);
 
+    //继承所有父类的成员变量和方法
+    if (superClass != null) {
+      for (final variable in superClass.variables.values) {
+        if (variable.isStatic) {
+          dynamic value;
+          if (variable.initializer != null) {
+            value = evaluateExpr(variable.initializer);
+          }
+          // else if (variable.isExtern) {
+          //   value = externs.fetch('${stmt.name}${HT_Lexicon.memberGet}${variable.name.lexeme}', variable.name.line,
+          //       variable.name.column, this,
+          //       from: externs.fullName);
+          // }
+
+          klass.define(variable.name.lexeme, this,
+              declType: variable.declType, line: variable.name.line, column: variable.name.column, value: value);
+        } else {
+          klass.addVariable(variable);
+        }
+      }
+    }
+
     var save = curContext;
     curContext = klass;
     for (final variable in stmt.variables) {
@@ -691,9 +731,9 @@ class Interpreter implements ExprVisitor, StmtVisitor {
     curContext = save;
 
     for (final method in stmt.methods) {
-      if (klass.contains(method.internalName)) {
-        throw HTErr_Defined(method.name, method.keyword.line, method.keyword.column, curFileName);
-      }
+      // if (klass.contains(method.internalName)) {
+      //   throw HTErr_Defined(method.name, method.keyword.line, method.keyword.column, curFileName);
+      // }
 
       HT_Function func;
       HT_External externFunc;