+++ lib/src/resolver.dart
@@ -12,18 +12,42 @@ enum _ClassType {
 /// 负责对语句列表进行分析，并生成变量作用域
 class Resolver implements ExprVisitor, StmtVisitor {
   final Interpreter interpreter;
+  final List<Stmt> statements;
+  final String fileName;
+  String _libName = HT_Lexicon.globals;
 
   /// 代码块列表，每个代码块包含一个字典：key：变量标识符，value：变量是否已初始化
   final _blocks = <Map<String, bool>>[];
 
   final _classes = <ClassDeclStmt>[];
-  var _funcs = <FuncDeclStmt>[];
+  final _funcs = <FuncDeclStmt>[];
 
-  String _curFileName;
   FuncStmtType _curFuncType;
   _ClassType _curClassType = _ClassType.none;
 
-  Resolver(this.interpreter);
+  Resolver(this.interpreter, this.statements, this.fileName);
+
+  List<Stmt> resolve({String libName = HT_Lexicon.globals}) {
+    _libName = libName;
+    if (_libName != HT_Lexicon.globals) _beginBlock();
+
+    _beginBlock();
+    for (final stmt in statements) {
+      _resolveStmt(stmt);
+    }
+    for (final klass in _classes) {
+      _resolveClass(klass);
+    }
+    for (final func in _funcs) {
+      _resolveFunction(func);
+    }
+    _endBlock();
+    if ((libName != null) && (libName != HT_Lexicon.globals)) {
+      _endBlock();
+    }
+
+    return statements;
+  }
 
   void _beginBlock() => _blocks.add(<String, bool>{});
   void _endBlock() => _blocks.removeLast();
@@ -33,7 +57,7 @@ class Resolver implements ExprVisitor, StmtVisitor {
       var block = _blocks.last;
 
       if (block.containsKey(name) && error) {
-        throw HTErr_Defined(name, line, column, _curFileName);
+        // throw HTErr_Defined(name, line, column, _curFileName);
       }
       block[name] = define;
     }
@@ -57,28 +81,6 @@ class Resolver implements ExprVisitor, StmtVisitor {
     // Not found. Assume it is global.
   }
 
-  void resolve(List<Stmt> statements, {String fileName, String libName}) {
-    libName ??= HT_Lexicon.globals;
-    if (libName != HT_Lexicon.globals) _beginBlock();
-
-    _curFileName = fileName;
-
-    _beginBlock();
-    for (final stmt in statements) {
-      _resolveStmt(stmt);
-    }
-    for (final klass in _classes) {
-      _resolveClass(klass);
-    }
-    for (final func in _funcs) {
-      _resolveFunction(func);
-    }
-    _endBlock();
-    if ((libName != null) && (libName != HT_Lexicon.globals)) {
-      _endBlock();
-    }
-  }
-
   void _resolveBlock(List<Stmt> statements) {
     for (final stmt in statements) {
       _resolveStmt(stmt);
@@ -94,7 +96,7 @@ class Resolver implements ExprVisitor, StmtVisitor {
 
     _beginBlock();
     if (stmt.arity == -1) {
-      //_declare(env.lexicon.Arguments, stmt.keyword.line, stmt.keyword.column, define: true);
+      _declare(stmt.params.first.name.lexeme, stmt.keyword.line, stmt.keyword.column, define: true);
     } else {
       for (final param in stmt.params) {
         _declare(param.name.lexeme, param.name.line, param.name.column, define: true);
@@ -107,72 +109,88 @@ class Resolver implements ExprVisitor, StmtVisitor {
 
   void _resolveClass(ClassDeclStmt stmt) {
     final savedClassType = _curClassType;
-
-    if (stmt.superClass != null) {
-      if (stmt.name == stmt.superClass?.name?.lexeme) {
-        throw HTErr_Unexpected(stmt.superClass.toString(), stmt.keyword.line, stmt.keyword.column, _curFileName);
-      }
-      _resolveExpr(stmt.superClass);
-    }
+    // TODO: super表达式
     _blocks.last[HT_Lexicon.SUPER] = true;
 
     _curClassType = _ClassType.normal;
 
     _beginBlock();
-    for (final variable in stmt.variables) {
-      if (variable.isStatic) {
-        visitVarDeclStmt(variable);
-      }
-    }
 
-    final savedFuncList = _funcs;
-    _funcs = <FuncDeclStmt>[];
-    // 类静态函数，先注册函数名
-    for (final method in stmt.methods) {
-      if (method.isStatic) {
-        _declare(method.internalName, method.keyword.line, method.keyword.column, define: true);
-        if ((method.internalName.startsWith(HT_Lexicon.getter) || method.internalName.startsWith(HT_Lexicon.setter)) &&
-            !_blocks.last.containsKey(method.name)) {
-          _declare(method.name, method.keyword.line, method.keyword.column, define: true);
-        }
-        if (method.funcType != FuncStmtType.constructor) {
-          _funcs.add(method);
+    // 递归获取所有父类的静态变量和静态函数
+    var static_var_stmt = <VarDeclStmt>[];
+    var static_func_stmt = <FuncDeclStmt>[];
+    var cur_stmt = stmt;
+    while (cur_stmt != null) {
+      for (final varStmt in cur_stmt.variables) {
+        if (varStmt.isStatic) {
+          static_var_stmt.add(varStmt);
         }
       }
-    }
-    // 然后再解析函数定义
-    for (final stmt in _funcs) {
-      _resolveFunction(stmt);
-    }
 
-    _funcs = <FuncDeclStmt>[];
-    _beginBlock();
-    // 注册实例中的成员变量
-    _blocks.last[HT_Lexicon.THIS] = true;
-    for (final variable in stmt.variables) {
-      if (!variable.isStatic) {
-        visitVarDeclStmt(variable);
-      }
-    }
-    // 成员函数，先注册函数名
-    for (final method in stmt.methods) {
-      if (!method.isStatic) {
-        if (method.funcType != FuncStmtType.constructor) {
-          _declare(method.internalName, method.keyword.line, method.keyword.column, define: true);
-          if ((method.internalName.startsWith(HT_Lexicon.getter) ||
-                  method.internalName.startsWith(HT_Lexicon.setter)) &&
-              !_blocks.last.containsKey(method.name)) {
-            _declare(method.name, method.keyword.line, method.keyword.column, define: true);
+      for (final funcStmt in cur_stmt.methods) {
+        if (funcStmt.isStatic) {
+          if ((funcStmt.internalName.startsWith(HT_Lexicon.getter) ||
+                  funcStmt.internalName.startsWith(HT_Lexicon.setter)) &&
+              !_blocks.last.containsKey(funcStmt.name)) {
+            _declare(funcStmt.name, funcStmt.keyword.line, funcStmt.keyword.column, define: true);
+          } else {
+            _declare(funcStmt.internalName, funcStmt.keyword.line, funcStmt.keyword.column, define: true);
+          }
+          if (funcStmt.funcType != FuncStmtType.constructor) {
+            static_func_stmt.add(funcStmt);
           }
         }
-        _funcs.add(method);
+      }
+
+      cur_stmt = cur_stmt.superClassDeclStmt;
+    }
+    // 注册变量名
+    for (final varStmt in static_var_stmt) {
+      visitVarDeclStmt(varStmt);
+    }
+    // 解析函数定义
+    for (final funcStmt in static_func_stmt) {
+      _resolveFunction(funcStmt);
+    }
+
+    _beginBlock();
+    _blocks.last[HT_Lexicon.THIS] = true;
+    // 递归获取所有父类的成员变量和成员函数
+    var instance_var_stmt = <VarDeclStmt>[];
+    var instance_func_stmt = <FuncDeclStmt>[];
+    cur_stmt = stmt;
+    while (cur_stmt != null) {
+      for (final varStmt in cur_stmt.variables) {
+        if (!varStmt.isStatic) {
+          instance_var_stmt.add(varStmt);
+        }
+      }
+
+      for (final funcStmt in cur_stmt.methods) {
+        if (!funcStmt.isStatic) {
+          _declare(funcStmt.internalName, funcStmt.keyword.line, funcStmt.keyword.column, define: true);
+          if ((funcStmt.internalName.startsWith(HT_Lexicon.getter) ||
+                  funcStmt.internalName.startsWith(HT_Lexicon.setter)) &&
+              !_blocks.last.containsKey(funcStmt.name)) {
+            _declare(funcStmt.name, funcStmt.keyword.line, funcStmt.keyword.column, define: true);
+          }
+          instance_func_stmt.add(funcStmt);
+        }
+      }
+
+      cur_stmt = cur_stmt.superClassDeclStmt;
+    }
+    // 注册变量名
+    for (final varStmt in instance_var_stmt) {
+      if (!varStmt.isStatic) {
+        visitVarDeclStmt(varStmt);
       }
     }
-    // 然后再解析函数定义
-    for (final stmt in _funcs) {
-      _resolveFunction(stmt);
+    // 解析函数定义
+    for (final funcStmt in instance_func_stmt) {
+      _resolveFunction(funcStmt);
     }
-    _funcs = savedFuncList;
+
     _endBlock();
 
     _endBlock();
@@ -206,13 +224,18 @@ class Resolver implements ExprVisitor, StmtVisitor {
     }
   }
 
+  @override
+  dynamic visitLiteralFunctionExpr(LiteralFunctionExpr expr) {
+    _resolveFunction(expr.funcStmt);
+  }
+
   // @override
   // dynamic visitTypeExpr(TypeExpr expr) {}
 
   @override
   dynamic visitSymbolExpr(SymbolExpr expr) {
     if (_blocks.isNotEmpty && _blocks.last[expr.name] == false) {
-      throw HTErr_Initialized(expr.name.lexeme, expr.line, expr.column, _curFileName);
+      throw HTErr_Initialized(expr.name.lexeme, expr.line, expr.column, fileName);
     }
 
     _lookUpVar(expr, expr.name.lexeme);
@@ -233,7 +256,7 @@ class Resolver implements ExprVisitor, StmtVisitor {
   dynamic visitCallExpr(CallExpr expr) {
     _resolveExpr(expr.callee);
 
-    for (final arg in expr.args) {
+    for (final arg in expr.positionedArgs) {
       _resolveExpr(arg);
     }
   }
@@ -294,7 +317,7 @@ class Resolver implements ExprVisitor, StmtVisitor {
   @override
   void visitReturnStmt(ReturnStmt stmt) {
     if ((_curFuncType == null) || (_curFuncType == FuncStmtType.constructor)) {
-      throw HTErr_Unexpected(stmt.keyword.lexeme, stmt.keyword.line, stmt.keyword.column, _curFileName);
+      throw HTErr_Unexpected(stmt.keyword.lexeme, stmt.keyword.line, stmt.keyword.column, fileName);
     }
 
     if (stmt.expr != null) {
@@ -326,7 +349,7 @@ class Resolver implements ExprVisitor, StmtVisitor {
   @override
   void visitThisExpr(ThisExpr expr) {
     if (_curClassType == _ClassType.none) {
-      throw HTErr_Unexpected(expr.keyword.lexeme, expr.line, expr.column, _curFileName);
+      throw HTErr_Unexpected(expr.keyword.lexeme, expr.line, expr.column, fileName);
     }
 
     _lookUpVar(expr, expr.keyword.lexeme);