+++ test/interpreter_test.dart
@@ -3,31 +3,31 @@ import 'package:hetu_script/hetu_script.dart';
 import 'package:hetu_script/src/lexer.dart';
 
 void main() async {
+  var itp = await Hetu.init();
   group('lexer test -', () {
     test('lexer', () async {
-      final lexer = Lexer();
+      final lexer = Lexer(itp, '''
+                // this is a comment
+                var _Words: String = "hello world"
+                let n_42 = 42
+                void main() {
+                print(_Words);
+                }
+                ''');
       expect(
-        lexer
-            .lex('// this is a comment\n'
-                'var _Words: String = "hello world"\n'
-                'let n_42 = 42\n'
-                'void main() {\n'
-                'print(_Words);\n'
-                '}')
-            .toString(),
+        lexer.tokens.toString(),
         '[var, _Words, :, String, =, "hello world", let, n_42, =, 42, void, main, (, ), {, print, (, _Words, ), ;, }]',
       );
     });
   });
   group('interpreter error handling test -', () {
     test('const definition', () async {
-      var itp = await Hetu.init();
       expect(
         () {
-          itp.eval(
-              'let i = 42\n'
-              'i = 137',
-              style: ParseStyle.function);
+          itp.eval('''
+              let i = 42
+              i = 137
+              ''', style: ParseStyle.function);
         },
         throwsA(isA<HTErr_Mutable>()),
       );